server-validation-out

+++ app/Http/Controllers/NacteResultController.php
@@ -9,6 +9,7 @@ use App\Domain\Application\Models\OutResultDetail;
 use App\Domain\Application\Models\NacteResult;
 use App\Domain\Application\Models\Applicant;
 use Illuminate\Support\Facades\Http;
+use App\Domain\Application\Models\ApplicantProgramSelection;
 
 class NacteResultController extends Controller
 {
@@ -17,6 +18,10 @@ class NacteResultController extends Controller
      */
     public function confirm(Request $request)
     {
+        if(ApplicantProgramSelection::where('applicant_id',$request->get('applicant_id'))->count() != 0){
+            return redirect()->back()->with('error','The action cannot be performed at the moment'); 
+        }
+
         $detail = NacteResultDetail::find($request->get('nacte_result_detail_id'));
         $detail->verified = 1;
         $detail->save();
@@ -56,7 +61,9 @@ class NacteResultController extends Controller
 
     public function confirmNacteRegNumber(Request $request)
     {
-        
+        if(ApplicantProgramSelection::where('applicant_id',$request->get('applicant_id'))->count() != 0){
+            return redirect()->back()->with('error','The action cannot be performed at the moment'); 
+        }
 
         $nacte_detail = NacteResultDetail::find($request->get('nacte_result_detail_id'));
 
@@ -89,6 +96,10 @@ class NacteResultController extends Controller
 
     public function declineNacteRegNumber(Request $request)
     {
+        if(ApplicantProgramSelection::where('applicant_id',$request->get('applicant_id'))->count() != 0){
+            return redirect()->back()->with('error','The action cannot be performed at the moment'); 
+        }
+
         $detail = NacteResultDetail::find($request->get('nacte_result_detail_id'));
         if($detail->verified != 1){
     	    $detail->delete();
@@ -101,6 +112,10 @@ class NacteResultController extends Controller
      */
     public function nullifyNacteReg(Request $request)
     {
+        if(ApplicantProgramSelection::where('applicant_id',$request->get('applicant_id'))->count() != 0){
+            return redirect()->back()->with('error','The action cannot be performed at the moment'); 
+        }
+
         $detail = NacteResultDetail::find($request->get('detail_id'));
         $detail->delete();
         return redirect()->back()->with('error','NECTA results names do not match your application names');
@@ -111,6 +126,10 @@ class NacteResultController extends Controller
      */
     public function destroy(Request $request)
     {
+        if(ApplicantProgramSelection::where('applicant_id',$request->get('applicant_id'))->count() != 0){
+            return redirect()->back()->with('error','The action cannot be performed at the moment'); 
+        }
+
     	$detail = NacteResultDetail::find($request->get('nacte_result_detail_id'));
         if($detail->verified != 1){
     	NacteResult::where('nacte_result_detail_id',$request->get('nacte_result_detail_id'))->delete();
@@ -125,6 +144,10 @@ class NacteResultController extends Controller
      */
     public function nullify(Request $request)
     {
+        if(ApplicantProgramSelection::where('applicant_id',$request->get('applicant_id'))->count() != 0){
+            return redirect()->back()->with('error','The action cannot be performed at the moment'); 
+        }
+        
         $detail = NacteResultDetail::find($request->get('detail_id'));
         NacteResult::where('nacte_result_detail_id',$request->get('detail_id'))->delete();
         // $detail->results->delete();
